+++ apps/spaces/src/components/modals/QuizLibraryPreviewModal/index.tsx
@@ -151,7 +151,7 @@ export const QuizLibraryPreviewModal: FunctionComponent<Props> = ({
                 <ActionsDivider />
 
                 <Button
-                  text={t("dashboard.use_template_button")}
+                  text={t("dashboard.use_a_quiz_template_button")}
                   type="primary"
                   color={defaultTheme.colors.green7}
                   leftIcon={<FaCirclePlus size={16} />}
